Fix double tap issue on tabs
I still don't know what was causing this really but tapping on a tab
that altered its state when it appeared caused the app to return to the
previous tab immediately. Tapping on the desired tab a second time would
load it correctly. By setting the tab selection variable myself, it now
works for some reason.